import slack data from web page
