+++ app/public/src/css/configure-window.less
@@ -33,6 +33,23 @@
   z-index: 900;
   opacity: .4;
 }
+
+#import-slack-data-form {
+  .form-section {
+    margin: 10px 0;
+    .submit-button {
+      background: #2ab27b;
+      border-radius: 5px;
+      border: 0;
+      color: #fff;
+      line-height: 1.2rem;
+      font-size: 14px;
+      padding: 5px 10px;
+      text-shadow: 0 1px 1px rgba(0,0,0,.1);
+    }
+  }
+}
+
 #logger-configure-form {
   .clearfix;
   margin: 5px 0 0 0;
